refactor(load): move pkg/build to pkg/buildx/build like other vendored code
